Add decoder (sacd) for dsf files, allowing decoder strings without %o
Problem: sacd takes the name of the input file (-i) and generates a wav
file with the same base name, in other words, it doesn't allow you to
choose the name of the output file. pacpl assumes that the wav filename
will be "basename.$$.wav", so it fails when the decoder generates wav
files named "basename.wav".
Solution: We check the decoder string: if it contains %o, then we'll use
"basename.$$.wav" as the wav filename; if it doesn't contain %o, we'll
use "basename.wav" as the wav filename. This lets sacd to work properly,
and any other decoder that doesn't allow you to choose the name of the
output file.