Fix: Make shell-safe all filenames used in system() calls
Problem: system("string") runs string through the shell, so any
filenames within string have to be shell safe. convert() is quoting \
and $ and enclosing the filenames in double quotes, but this fails
whenever the filenames have embedded ` or ". There were other blocks of
code with system() that don't quote anything at all in the filenames.
Solution: Use method shell_quote from module String::ShellQuote, which
encloses a string in single quotes and escapes it so that it's shell
safe. ALL system() calls which handled filenames were changed: the
filenames were replaced by the result of "shell_quote filename".